requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.0
pandas==2.1.1
numpy==1.24.3 